Add module with models
It contains the function runge_kutta() which implements
the 4th order Runge-Kutta method for solving
the system of fist order differential equations
with Cauchy initial conditions.

The function runge_kutta_tolerance() implements
the the 4th order Runge-Kutta scheme
and compares it with the Euler method.
If the differences do not fit into the tolerance,
then the schema increment step is changed.

The function rhs() calculates the array of
right hand side functions of the system of equations.